Event 2023 Day 21 - now three solution, all wrong

+++ AdventofCode/Source/2023 Solutions/AoC2023_24.h
@@ -244,7 +244,7 @@ template<> Number AoC<2023, 24, B>(std::istream& input)
       return r;
     };
   
-  int s = std::ssize(hail);
+  int s = static_cast<int>(std::ssize(hail));
   Vector3D<Real> rock{};
   for (int i = 0; i < s; i++)
   {
